api/core/celery.py
import os
from celery import Celery
from dotenv import load_dotenv

load_dotenv()

celery_app = Celery(
    'crew_tasks',
    broker=os.getenv("CELERY_BROKER_URL"),
    backend=os.getenv("CELERY_RESULT_BACKEND"),
    include=['celery_worker']
)