requirements.txt
crewai
crewai_tools
fastapi
uvicorn
celery
redis
requests
python-dotenv